refactor: extract forge-protocol so the relay skips the engine
forge-server (the relay) only needs the wire types (ClientMessage,
ServerMessage, RoomInfo, Deck) but depended on forge-agent-interface,
which pulls in forge-engine-core -> forge-carddb -> the cardset archive
via its engine-coupled DTOs (prompts, game views, StateEnvelope) — none
of which the relay uses (it routes in-game payloads as opaque Value).

Move the engine-free wire types + deck_dto into a new forge-protocol
crate (deps: serde only). forge-agent-interface re-exports them so all
existing `forge_agent_interface::protocol::*` paths keep working, and
keeps StateEnvelope (engine-coupled). forge-server and self-hosted-node
now source the protocol from forge-protocol / forge-agent-interface
directly, so the relay's closure is just {forge-server, forge-protocol}
and it no longer compiles the engine.

deploy.sh's forge-server trigger collapses to that two-crate closure.
