+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "forge-engine/crates/forge-carddb",
     "forge-engine/crates/forge-engine-macros",
     "forge-engine/crates/forge-engine",
+    "forge-engine/crates/forge-protocol",
     "forge-engine/crates/forge-server",
     "forge-engine/crates/self-hosted-node",
     "forge-engine/crates/forge-game-runtime",
