+++ forge-engine/crates/forge-protocol/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "forge-protocol"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { workspace = true }
+serde_json = { workspace = true }
